requirements.txt
kafka-python
pandas
orjson
psycopg2-binary
